Refactor controller, add ProducesResponseType and fix projects dependencies
