+++ AccessCorp.Identity/AccessCorp.Domain/AccessCorp.Domain.csproj
@@ -11,4 +11,8 @@
         <Folder Include="Entities\" />
     </ItemGroup>
 
+    <ItemGroup>
+      <PackageReference Include="Microsoft.AspNetCore.Identity.UI" Version="8.0.10" />
+    </ItemGroup>
+
 </Project>
