+++ AccessCorp.Identity/AccessCorp.WebApi/Program.cs
@@ -15,6 +15,7 @@ builder.AddApiConfiguration()
 builder.Services.AddScoped<IAuthService, AuthService>();
 builder.Services.AddScoped<IUserClaimsService, UserClaimsService>();
 builder.Services.AddScoped<ICepValidationService, CepValidationService>();
+builder.Services.AddScoped<IJwtService, JwtService>();
 
 builder.Services.AddSingleton<IAuthorizationHandler, PermissionHandler>();
